Refactor DataFoldNode modules
