+++ fold_node/src/datafold_node/mod.rs
@@ -13,6 +13,9 @@ pub mod config;
 pub mod error;
 pub mod http_server;
 pub mod loader;
+mod db;
+mod permissions;
+mod transform_queue;
 pub mod node;
 pub mod tcp_server;
 pub mod tests;
