+++ fold_node/src/datafold_node/node.rs
@@ -9,8 +9,6 @@ use crate::datafold_node::config::NodeInfo;
 use crate::error::{FoldDbError, FoldDbResult, NetworkErrorKind};
 use crate::fold_db_core::FoldDB;
 use crate::network::{NetworkConfig, NetworkCore, PeerId};
-use crate::schema::types::{Mutation, Operation, Query, Transform};
-use crate::schema::{Schema, SchemaError};
 
 /// A node in the DataFold distributed database system.
 ///
@@ -59,15 +57,15 @@ use crate::schema::{Schema, SchemaError};
 #[derive(Clone)]
 pub struct DataFoldNode {
     /// The underlying database instance for data storage and operations
-    db: Arc<Mutex<FoldDB>>,
+    pub(super) db: Arc<Mutex<FoldDB>>,
     /// Configuration settings for this node
-    config: NodeConfig,
+    pub(super) config: NodeConfig,
     /// Map of trusted nodes and their trust distances
-    trusted_nodes: HashMap<String, NodeInfo>,
+    pub(super) trusted_nodes: HashMap<String, NodeInfo>,
     /// Unique identifier for this node
-    node_id: String,
+    pub(super) node_id: String,
     /// Network layer for P2P communication
-    network: Option<Arc<tokio::sync::Mutex<NetworkCore>>>,
+    pub(super) network: Option<Arc<tokio::sync::Mutex<NetworkCore>>>,
 }
 
 /// Basic status information about the network layer
@@ -112,337 +110,6 @@ impl DataFoldNode {
         Self::new(config)
     }
 
-    /// Loads a schema into the database and grants this node permission.
-    ///
-    /// This function loads a schema into the database, making it available for
-    /// queries and mutations. It also grants the local node permission to access
-    /// the schema.
-    ///
-    /// # Arguments
-    ///
-    /// * `schema` - The schema to load
-    ///
-    /// # Returns
-    ///
-    /// A `FoldDbResult` indicating success or failure.
-    ///
-    /// # Errors
-    ///
-    /// Returns a `FoldDbError` if:
-    /// * There is an error loading the schema into the database
-    /// * There is an error granting permission to the local node
-    ///
-    /// # Examples
-    ///
-    /// ```rust,no_run
-    /// use fold_node::datafold_node::{DataFoldNode, NodeConfig};
-    /// use fold_node::schema::Schema;
-    /// use fold_node::error::FoldDbResult;
-    /// use std::path::PathBuf;
-    /// use std::collections::HashMap;
-    ///
-    /// fn main() -> FoldDbResult<()> {
-    ///     let config = NodeConfig::new(PathBuf::from("data"));
-    ///     let mut node = DataFoldNode::new(config)?;
-    ///     
-    ///     // Create a schema
-    ///     let schema = Schema::new("user_profile".to_string());
-    ///
-    ///     // Load the schema
-    ///     node.load_schema(schema)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    pub fn load_schema(&mut self, schema: Schema) -> FoldDbResult<()> {
-        let schema_name = schema.name.clone();
-        let mut db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        db.load_schema(schema)?;
-        drop(db);
-        self.grant_schema_permission(&schema_name)?;
-        Ok(())
-    }
-
-    /// Executes an operation (query or mutation) on the database.
-    ///
-    /// This function processes a schema operation, which can be either a query
-    /// to retrieve data or a mutation to modify data. It handles permission
-    /// checking, operation routing, and result formatting.
-    ///
-    /// # Arguments
-    ///
-    /// * `operation` - The operation to execute, either a Query or Mutation
-    ///
-    /// # Returns
-    ///
-    /// A `FoldDbResult` containing the JSON result of the operation.
-    /// For queries, this will be an array of values for the requested fields.
-    /// For mutations, this will typically be null or a success indicator.
-    ///
-    /// # Errors
-    ///
-    /// Returns a `FoldDbError` if:
-    /// * The schema does not exist
-    /// * The operation is not permitted
-    /// * There is an error executing the operation
-    /// * The result cannot be serialized
-    ///
-    /// # Examples
-    ///
-    /// ```rust,no_run
-    /// use fold_node::datafold_node::{DataFoldNode, NodeConfig};
-    /// use fold_node::schema::types::{Operation, MutationType};
-    /// use fold_node::error::FoldDbResult;
-    /// use std::path::PathBuf;
-    /// use serde_json::json;
-    ///
-    /// fn main() -> FoldDbResult<()> {
-    ///     let config = NodeConfig::new(PathBuf::from("data"));
-    ///     let mut node = DataFoldNode::new(config)?;
-    ///
-    ///     // Execute a query
-    ///     let query_op = Operation::Query {
-    ///         schema: "user_profile".to_string(),
-    ///         fields: vec!["username".to_string(), "email".to_string()],
-    ///         filter: None,
-    ///     };
-    ///     let query_result = node.execute_operation(query_op)?;
-    ///
-    ///     // Execute a mutation
-    ///     let mutation_op = Operation::Mutation {
-    ///         schema: "user_profile".to_string(),
-    ///         data: json!({
-    ///             "username": "new_user",
-    ///             "email": "user@example.com"
-    ///         }),
-    ///         mutation_type: MutationType::Create,
-    ///     };
-    ///     let mutation_result = node.execute_operation(mutation_op)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    pub fn execute_operation(&mut self, operation: Operation) -> FoldDbResult<Value> {
-        info!("Executing operation: {:?}", operation);
-        match operation {
-            Operation::Query {
-                schema,
-                fields,
-                filter: _,
-            } => {
-                let query = Query {
-                    schema_name: schema,
-                    fields,
-                    pub_key: String::new(), // TODO: Get from auth context
-                    trust_distance: 0,      // Set write distance to 0 for all queries
-                };
-
-                let results = self.query(query)?;
-
-                let unwrapped_results: Vec<Value> = results
-                    .into_iter()
-                    .map(|result| match result {
-                        Ok(value) => value,
-                        Err(e) => serde_json::json!({"error": e.to_string()}),
-                    })
-                    .collect();
-
-                Ok(serde_json::to_value(&unwrapped_results)?)
-            }
-            Operation::Mutation {
-                schema,
-                data,
-                mutation_type,
-            } => {
-                let fields_and_values = match data {
-                    Value::Object(map) => map.into_iter().collect(),
-                    _ => {
-                        return Err(FoldDbError::Config(
-                            "Mutation data must be an object".into(),
-                        ))
-                    }
-                };
-
-                info!("Mutation type: {:?}", mutation_type);
-
-                let mutation = Mutation {
-                    schema_name: schema,
-                    fields_and_values,
-                    pub_key: String::new(), // TODO: Get from auth context
-                    trust_distance: 0,      // Set write distance to 0 for all mutations
-                    mutation_type,
-                };
-
-                self.mutate(mutation)?;
-
-                Ok(Value::Null)
-            }
-        }
-    }
-
-    /// Retrieves a schema by its ID.
-    pub fn get_schema(&self, schema_id: &str) -> FoldDbResult<Option<Schema>> {
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        Ok(db.schema_manager.get_schema(schema_id)?)
-    }
-
-    /// Lists all loaded schemas in the database.
-    pub fn list_schemas(&self) -> FoldDbResult<Vec<Schema>> {
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        let schema_names = db.schema_manager.list_schemas()?;
-        // info!("Schema names from schema_manager: {:?}", schema_names);
-        let mut schemas = Vec::new();
-        for name in schema_names {
-            if let Some(schema) = db.schema_manager.get_schema(&name)? {
-                schemas.push(schema);
-            }
-        }
-        Ok(schemas)
-    }
-
-    /// Executes a query against the database.
-    pub fn query(&self, mut query: Query) -> FoldDbResult<Vec<Result<Value, SchemaError>>> {
-        // Enforce per-node schema permission
-        if !self.check_schema_permission(&query.schema_name)? {
-            return Err(FoldDbError::Config(format!(
-                "Permission denied for schema {}",
-                query.schema_name
-            )));
-        }
-        if query.trust_distance == 0 {
-            query.trust_distance = self.config.default_trust_distance;
-        }
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        Ok(db.query_schema(query))
-    }
-
-    /// Executes a mutation on the database.
-    pub fn mutate(&mut self, mutation: Mutation) -> FoldDbResult<()> {
-        // Enforce per-node schema permission
-        if !self.check_schema_permission(&mutation.schema_name)? {
-            return Err(FoldDbError::Config(format!(
-                "Permission denied for schema {}",
-                mutation.schema_name
-            )));
-        }
-        let mut db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        db.write_schema(mutation)?;
-        Ok(())
-    }
-
-    /// Adds a trusted node to the node's trusted nodes list.
-    pub fn add_trusted_node(&mut self, node_id: &str) -> FoldDbResult<()> {
-        self.trusted_nodes.insert(
-            node_id.to_string(),
-            NodeInfo {
-                id: node_id.to_string(),
-                trust_distance: self.config.default_trust_distance,
-            },
-        );
-        Ok(())
-    }
-
-    /// Removes a trusted node from the node's trusted nodes list.
-    pub fn remove_trusted_node(&mut self, node_id: &str) -> FoldDbResult<()> {
-        self.trusted_nodes.remove(node_id);
-        Ok(())
-    }
-
-    /// Gets the current list of trusted nodes and their trust distances.
-    pub fn get_trusted_nodes(&self) -> &HashMap<String, NodeInfo> {
-        &self.trusted_nodes
-    }
-
-    /// Retrieves the version history for a specific atom reference.
-    pub fn get_history(&self, aref_uuid: &str) -> FoldDbResult<Vec<Value>> {
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        let history = db
-            .atom_manager
-            .get_atom_history(aref_uuid)
-            .map_err(|e| FoldDbError::Database(e.to_string()))?;
-
-        Ok(history
-            .into_iter()
-            .map(|atom| atom.content().clone())
-            .collect())
-    }
-
-    /// Allows operations on a schema and persists permission for this node.
-    pub fn allow_schema(&mut self, schema_name: &str) -> FoldDbResult<()> {
-        let mut db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        db.allow_schema(schema_name)?;
-        // Persist this node's permission
-        drop(db); // release lock before acquiring in grant_schema_permission
-        self.grant_schema_permission(schema_name)?;
-        Ok(())
-    }
-
-    /// Grants schema permission for this node.
-    pub fn grant_schema_permission(&mut self, schema_name: &str) -> FoldDbResult<()> {
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        let mut perms = db.get_schema_permissions(&self.node_id);
-        if !perms.contains(&schema_name.to_string()) {
-            perms.push(schema_name.to_string());
-            db.set_schema_permissions(&self.node_id, &perms)
-                .map_err(|e| FoldDbError::Config(format!("Failed to set permissions: {}", e)))?;
-        }
-        Ok(())
-    }
-
-    /// Revokes schema permission for this node.
-    pub fn revoke_schema_permission(&mut self, schema_name: &str) -> FoldDbResult<()> {
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        let mut perms = db.get_schema_permissions(&self.node_id);
-        perms.retain(|s| s != schema_name);
-        db.set_schema_permissions(&self.node_id, &perms)
-            .map_err(|e| FoldDbError::Config(format!("Failed to set permissions: {}", e)))?;
-        Ok(())
-    }
-
-    /// Checks if this node has permission to access the given schema.
-    pub fn check_schema_permission(&self, schema_name: &str) -> FoldDbResult<bool> {
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        let perms = db.get_schema_permissions(&self.node_id);
-        Ok(perms.contains(&schema_name.to_string()))
-    }
-
-    /// Removes a schema from the database.
-    pub fn remove_schema(&mut self, schema_name: &str) -> FoldDbResult<()> {
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-
-        db.remove_schema(schema_name).map_err(|e| e.into())
-    }
 
     /// Initialize the network layer
     pub async fn init_network(&mut self, network_config: NetworkConfig) -> FoldDbResult<()> {
@@ -681,64 +348,9 @@ impl DataFoldNode {
         })
     }
 
-    /// List all registered transforms.
-    pub fn list_transforms(&self) -> FoldDbResult<HashMap<String, Transform>> {
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        Ok(db.list_transforms()?)
-    }
-
-    /// Execute a transform by id and return the result.
-    pub fn run_transform(&mut self, transform_id: &str) -> FoldDbResult<Value> {
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        Ok(db.run_transform(transform_id)?)
-    }
-
     /// Gets the unique identifier for this node.
     pub fn get_node_id(&self) -> &str {
         &self.node_id
     }
 
-    /// Add a transform to the queue
-    pub fn add_transform_to_queue(&self, transform_id: &str) -> FoldDbResult<()> {
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        
-        db.transform_orchestrator.add_transform(transform_id)?;
-        Ok(())
-    }
-
-    /// Get information about the transform queue
-    pub fn get_transform_queue_info(&self) -> FoldDbResult<serde_json::Value> {
-        let db = self
-            .db
-            .lock()
-            .map_err(|_| FoldDbError::Config("Cannot lock database mutex".into()))?;
-        
-        let queue_length = db.transform_orchestrator.len()?;
-        let is_empty = db.transform_orchestrator.is_empty()?;
-        
-        // Get the queued transform IDs
-        let mut queue = Vec::new();
-        let mut current_length = queue_length;
-        while current_length > 0 {
-            if let Some(Ok(transform_id)) = db.transform_orchestrator.process_one() {
-                queue.push(transform_id.to_string());
-            }
-            current_length -= 1;
-        }
-
-        Ok(serde_json::json!({
-            "queue": queue,
-            "length": queue_length,
-            "isEmpty": is_empty
-        }))
-    }
 }
